# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agency3 docs",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://tender.jti.jti.t-agency.ru/"
        }
    ],
    "paths": {
...
        "/api/1.0/user/auth": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Авторизация по логину и паролю",
                "description": "",
                "operationId": "userAuth",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test.email@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "test.email@example.com",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь авторизован, json {'message':''}"
                    },
                    "401": {
                        "description": "Логин или пароль не верный, json {'message':''}"
                    }
                }
            }
        },
...
    }
}